refactor: move odp relay macro into odp-client; delete embedded-service relay module
