+++ espi-service/Cargo.toml
@@ -21,6 +21,7 @@ embassy-sync.workspace = true
 embassy-imxrt = { workspace = true, features = ["mimxrt633s"] }
 embassy-futures.workspace = true
 mctp-rs = { workspace = true, features = ["espi"] }
+odp-client.workspace = true
 odp-service-common.workspace = true
 
 [target.'cfg(target_os = "none")'.dependencies]
@@ -43,6 +44,7 @@ defmt = [
     "embassy-sync/defmt",
     "embassy-imxrt/defmt",
     "mctp-rs/defmt",
+    "odp-client/defmt",
 ]
 
 log = ["dep:log", "embedded-services/log"]
