+++ examples/rt685s-evk/Cargo.toml
@@ -50,6 +50,7 @@ mimxrt600-fcb = "0.1.0"
 
 embedded-cfu-protocol = { git = "https://github.com/OpenDevicePartnership/embedded-cfu"}
 embedded-services = { path = "../../embedded-service", features = ["defmt"] }
+odp-client = { path = "../../odp-client", features = ["defmt"] }
 odp-service-common = { path = "../../odp-service-common" }
 power-button-service = { path = "../../power-button-service", features = [
     "defmt",
