+++ thermal-service-relay/Cargo.toml
@@ -8,6 +8,7 @@ repository.workspace = true
 [dependencies]
 defmt = { workspace = true, optional = true }
 embedded-services.workspace = true
+odp-client.workspace = true
 thermal-service-interface.workspace = true
 num_enum.workspace = true
 uuid.workspace = true
@@ -16,4 +17,4 @@ uuid.workspace = true
 workspace = true
 
 [features]
-defmt = ["dep:defmt"]
+defmt = ["dep:defmt", "odp-client/defmt"]
